[update] axios call replaced with data handler call [refactor] moved to directory package
